% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spectra.R
\name{add_spectra_column}
\alias{add_spectra_column}
\title{Add spectra column to metadata}
\usage{
add_spectra_column(data, specdb, ...)
}
\arguments{
\item{data}{`data.frame` to which to add spectra. Must include columns 
`project_code` and `observation_id`}

\item{specdb}{Path to spectra database}

\item{...}{Additional arguments to [load_spectra]}
}
\value{
`data`, with `spectra` as a list column of [PEcAnRTM::spectra] objects.
}
\description{
Add spectra column to metadata
}
