% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_helpers.R
\name{as.H5File}
\alias{as.H5File}
\alias{as.H5File.character}
\alias{as.H5File.H5File}
\title{Coerce object into H5File}
\usage{
as.H5File(x, ...)

\method{as.H5File}{character}(x, ...)

\method{as.H5File}{H5File}(x, ...)
}
\arguments{
\item{x}{Object to coerce to H5File}
}
\value{
H5File object
}
\description{
For characters, open an H5File object. For `H5File`s, return the object 
itself.
}
