% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_iff.R
\name{censor_if}
\alias{censor_if}
\title{Set value to NA if it matches some condition}
\usage{
censor_if(x, cond)
}
\arguments{
\item{x}{Value to censor}

\item{cond}{Logical vector of conditions at which to censor}
}
\description{
Set value to NA if it matches some condition
}
