% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_legend.R
\name{create_legend}
\alias{create_legend}
\title{Create legend based on a named color pallete}
\usage{
create_legend(color_pallete, leg_title, theme = ggplot2::theme())
}
\arguments{
\item{color_pallete}{Named color pallete}

\item{leg_title}{Legend title}

\item{theme}{ggplot2 theme object with which to modify legend}
}
\description{
Create legend based on a named color pallete
}
