% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create a project in the spectra database}
\usage{
create_project(specdb, project_code, spectra, metadata, overwrite = FALSE)
}
\arguments{
\item{specdb}{Path to spectra database directory}

\item{project_code}{Character string of project code}

\item{spectra}{Spectra object}
}
\description{
Create a project in the spectra database
}
