% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_specfile.R
\name{create_specfile}
\alias{create_specfile}
\title{Create the spectra NetCDF file}
\usage{
create_specfile(filename)
}
\arguments{
\item{filename}{Name of spectra database}
}
\description{
Create the spectra NetCDF file
}
