% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{df2dict}
\alias{df2dict}
\title{Convert two columns of a data frame to a named vector}
\usage{
df2dict(.data, values = 2, keys = 1)
}
\arguments{
\item{.data}{Data frame from which to extract names}

\item{values}{Column containing dictionary values}

\item{keys}{Column containing dictionary keys}
}
\description{
Convert two columns of a data frame to a named vector
}
