% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spectra.R
\name{get_metadata}
\alias{get_metadata}
\title{Retrieve all metadata for given project}
\usage{
get_metadata(specdb, projects = NULL)
}
\arguments{
\item{specdb}{H5File object containing spectra database, or character path 
to containing file}

\item{projects}{Character vector of projects for which to load metadata. If 
`NULL` (default), load all projects.}
}
\value{
Metadata for specified projects, as `tibble`
}
\description{
Retrieve all metadata for given project
}
