% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_project_metadata}
\alias{get_project_metadata}
\title{Retrieve metadata for an individual project}
\usage{
get_project_metadata(specdb, project, unfactor = FALSE,
  date_as_posix = FALSE, metadata_cols = c("project_code", "short_name"),
  pb = NULL)
}
\arguments{
\item{specdb}{Path to spectra database}

\item{project}{String containing project name}

\item{unfactor}{Logical. If `TRUE`, convert factor columns to character}

\item{date_as_posix}{Convert `Date` columns to `POSIXct`}

\item{metadata_cols}{Vector of metadata items to extract as columns}

\item{pb}{Progress bar object (default = `NULL`)}
}
\description{
Retrieve metadata for an individual project
}
