% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spectra.R
\name{get_spectra}
\alias{get_spectra}
\title{Retrieve spectra from NetCDF database}
\usage{
get_spectra(spectra_id, filename, wavelengths = seq(400, 2500))
}
\arguments{
\item{spectra_id}{Vector of integer IDs of spectra to retrieve}

\item{filename}{Name of spectra database}

\item{wavelengths}{Vector of wavelengths for retrieved spectra}
}
\description{
Retrieve spectra from NetCDF database
}
