% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inversion.R
\name{get_status_table}
\alias{get_status_table}
\title{Build a table of information about status of current inversions}
\usage{
get_status_table(specdb, prospect_version, spectra_types = c("R", "PA",
  "CRR"))
}
\arguments{
\item{specdb}{Path to spectra database}

\item{prospect_version}{Version of PROSPECT to use. Default is "5B".}

\item{spectra_types}{Spectra types to retrieve, or `NULL` (default) for all 
types. See [PEcAnRTM::valid_spectra_types].}
}
\description{
Build a table of information about status of current inversions
}
