% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_helpers.R
\name{h5_dataset}
\alias{h5_dataset}
\alias{h5_dataset.H5RefClass}
\title{Open a dataset, or create it if it doesn't exist}
\usage{
h5_dataset(hfile, dataset, data = NULL, overwrite = FALSE)

\method{h5_dataset}{H5RefClass}(hfile, dataset, data = NULL,
  overwrite = FALSE)
}
\arguments{
\item{hfile}{H5File object, or path to HDF5 file.}

\item{dataset}{Name of target dataset, as character}

\item{data}{R object to store in dataset}

\item{overwrite}{Logical. If `TRUE`, unlink the target first.}
}
\description{
Open a dataset, or create it if it doesn't exist
}
