% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_helpers.R
\name{h5_group}
\alias{h5_group}
\title{Open a group, or create it if it doesn't exist}
\usage{
h5_group(hfile, group, overwrite = FALSE)
}
\arguments{
\item{hfile}{HDF5 file object (see [hdf5r::H5File]).}

\item{group}{Name of group, as character}

\item{overwrite}{Logical. If `TRUE`, unlink the target first.}
}
\value{
H5File object of the resulting group
}
\description{
Open a group, or create it if it doesn't exist
}
