% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_helpers.R
\name{h5_open}
\alias{h5_open}
\title{Try to open an HDF5 file}
\usage{
h5_open(filename, sleep = 0.2, tries = 100, ...)
}
\arguments{
\item{filename}{Path to HDF5 file}

\item{sleep}{Seconds to sleep between attempts (default = 0.2)}

\item{tries}{Number of attempts (default = 100)}

\item{...}{Additional arguments to `H5File$new`}
}
\description{
This is meant to try to circumvent issues with multiple R sessions opening 
the same HDF5 file.
}
