% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms2dd.R
\name{hms2dd}
\alias{hms2dd}
\title{Convert hour-minute-second latitude/longitude data to decimal degrees}
\usage{
hms2dd(h, m, s, hemi)
}
\arguments{
\item{h}{Hour coordinate}

\item{m}{Minute coordinate}

\item{s}{Second coordinate}

\item{hemi}{Hemisphere (e.g. "N", "E", "S", "W") as character}
}
\value{
Decimal degree values
}
\description{
Convert hour-minute-second latitude/longitude data to decimal degrees
}
