% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhere.R
\name{infile}
\alias{infile}
\alias{indir}
\title{Quickly create relative paths to files and directories}
\usage{
infile(...)

indir(...)
}
\arguments{
\item{...}{List of paths, as in [fs::path]}
}
\value{
File name, as character
}
\description{
Quickly create relative paths to files and directories
}
