% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_helpers.R
\name{list2hdf}
\alias{list2hdf}
\alias{list2hdf.H5RefClass}
\alias{list2hdf.character}
\title{Recursively insert a list into an HDF5 file}
\usage{
list2hdf(hfile, l, overwrite = FALSE)

\method{list2hdf}{H5RefClass}(hfile, l, overwrite = FALSE)

\method{list2hdf}{character}(hfile, l, overwrite = FALSE)
}
\arguments{
\item{hfile}{H5File object, or path to HDF5 file.}

\item{l}{A named list}

\item{overwrite}{Logical. If `TRUE`, unlink the target first.}
}
\description{
Recursively insert a list into an HDF5 file
}
