% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spectra.R
\name{load_spectra}
\alias{load_spectra}
\title{Load single set of observed spectra from HDF5 file}
\usage{
load_spectra(project, observation_id, specdb, wl_min = NULL, wl_max = NULL,
  wavelengths = NULL, spectra_types = NULL)
}
\arguments{
\item{project}{Character string of project code from which to load spectra}

\item{observation_id}{Observation ID of spectra to load}

\item{specdb}{H5File object containing spectra database, or character path 
to containing file}

\item{wl_min}{Minimum wavelength to retrieve, or `NULL` (default) for no minimum}

\item{wl_max}{Maximum wavelength to retrieve, or `NULL` (default) for no maximum}

\item{wavelengths}{Wavelengths to which to [PEcAnRTM::resample] results, or 
`NULL` (default) to retrieve as is}

\item{spectra_types}{Spectra types to retrieve, or `NULL` (default) for all 
types. See [PEcAnRTM::valid_spectra_types].}
}
\value{
[PEcAnRTM::spectra] object
}
\description{
Load single set of observed spectra from HDF5 file
}
