% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_long_spectra.R
\name{make_long_spectra}
\alias{make_long_spectra}
\title{Convert a `spectra` object into a long `tibble`}
\usage{
make_long_spectra(spectra)
}
\arguments{
\item{spectra}{Object of class `spectra`}
}
\value{
Long `tibble` with each wavelength and observation in one row
}
\description{
Also, create a unique `spectra_id` column, and change the column
names to match the specification in the README.
}
