% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_spectra.R
\name{mat2list}
\alias{mat2list}
\title{Split a matrix-like object into a list of matrices based on the groups}
\usage{
mat2list(mat, groupvec, group = unique(groupvec))
}
\description{
Split a matrix-like object into a list of matrices based on the groups
}
