% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spectra.R
\name{pull_spectra}
\alias{pull_spectra}
\title{Convert spectra list column to single spectra object}
\usage{
pull_spectra(data, spectra_col = "spectra")
}
\arguments{
\item{data}{`data.frame` from which to extract spectra.}

\item{spectra_col}{Column name (quoted or unquoted, as in [dplyr::pull]) containing spectra}
}
\value{
Single [PEcAnRTM::spectra] object containing all spectra in `data`
}
\description{
Convert spectra list column to single spectra object
}
