% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spectra.R
\name{read_spectra}
\alias{read_spectra}
\title{Read spectra from CSVY file}
\usage{
read_spectra(filename, ...)
}
\arguments{
\item{filename}{Target file to read}

\item{...}{Additional arguments to [metar::read_csvy]}
}
\value{
Object of class `spectra`
}
\description{
Read spectra from CSVY file
}
