% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inversion.R
\name{run_inversion}
\alias{run_inversion}
\title{Invert a spectrum from the spectra database}
\usage{
run_inversion(project, observation_id, specdb, prospect_version,
  spectra_types = c("R", "PA", "CRR"), overwrite = FALSE, ...)
}
\arguments{
\item{project}{Character string of project code from which to load spectra}

\item{observation_id}{Observation ID of spectra to load}

\item{specdb}{H5File object containing spectra database, or character path 
to containing file}

\item{prospect_version}{Version of PROSPECT to use. Default is "5B".}

\item{spectra_types}{Spectra types to retrieve, or `NULL` (default) for all 
types. See [PEcAnRTM::valid_spectra_types].}

\item{overwrite}{Logical. If `TRUE`, overwrite all information related to 
this run. Default = `FALSE.`}

\item{...}{Additional parameters to [PEcAnRTM::invert_fieldspec]}
}
\description{
Invert a spectrum from the spectra database
}
