% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_helpers.R
\name{sanitize_df}
\alias{sanitize_df}
\title{Sanitize data for HDF5 storage}
\usage{
sanitize_df(data)
}
\arguments{
\item{data}{`data.frame` or `list` to sanitize}
}
\description{
Sanitize data for HDF5 storage
}
