% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inversion.R
\name{set_inversion_status}
\alias{set_inversion_status}
\title{Set and retrieve the status of an inversion}
\usage{
set_inversion_status(specdb, project, observation_id, prospect_version, status)
}
\arguments{
\item{specdb}{Path to spectra database}

\item{observation_id}{Observation ID of spectra to load}

\item{prospect_version}{Version of PROSPECT to use. Default is "5B".}

\item{status}{Character indicating status}
}
\description{
Set and retrieve the status of an inversion
}
