% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specdb.R
\name{specdb}
\alias{specdb}
\title{Spectra database class}
\usage{
specdb(path)
}
\value{
Object of class `specdb`
}
\description{
A specialized file hierarchy for storing spectra and metadata.
}
\details{
The database is organized as follows.
 - Database root contains the following:
   - Folders for each project, named according to `project_code`
   - `species.csvy` -- Information about species
   - `instruments.yml` -- Information about instruments
   - `traits.yml` -- Information about traits and units
 - Each project folder contains the following:
   - `metadata.csvy` -- Metadata about the spectra
   - `spectra/` -- Directory containing spectr
 - Each `spectra` folder contains `<observation_id.csvy>` files
}
