% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{swap_names}
\alias{swap_names}
\title{Swap the keys and values of a named vector}
\usage{
swap_names(dict)
}
\arguments{
\item{dict}{A named vector}
}
\description{
Swap the keys and values of a named vector
}
