% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_df.R
\docType{data}
\name{variable_df}
\alias{variable_df}
\title{Data frame of PROSPECT parameters}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.}
\usage{
variable_df
}
\description{
Data frame of PROSPECT parameters
}
\keyword{datasets}
