% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spectra.R
\name{write_spectra}
\alias{write_spectra}
\title{Write spectra to HDF5 file}
\usage{
write_spectra(spectra, data_name, spectra_file, overwrite = FALSE)
}
\arguments{
\item{spectra}{Spectra to store, either a matrix with wavelengths as row 
names, or an object of class `spectra`}

\item{data_name}{Data name for storing spectra}

\item{spectra_file}{File name for HDF5 file}

\item{overwrite}{Delete dataset `data_name` in HDF5 file before writing.}
}
\value{
Spectra file name (invisibly)
}
\description{
Write spectra to HDF5 file
}
