% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create a project in the spectra database}
\usage{
create_project(specdb_file, project_code, project_info, spectra, metadata,
  overwrite = FALSE)
}
\arguments{
\item{specdb_file}{Path to spectra file}

\item{project_code}{Character string of project code}

\item{project_info}{List of common project information}

\item{spectra}{Spectra object}
}
\description{
Create a project in the spectra database
}
