% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R, R/load_spectra.R
\name{get_metadata}
\alias{get_metadata}
\alias{get_metadata}
\title{Retrieve all metadata for given list of projects}
\usage{
get_metadata(specdb, projects = NULL, include_metadata = TRUE,
  unfactor = TRUE)

get_metadata(specdb, projects = NULL, include_metadata = TRUE,
  unfactor = TRUE)
}
\arguments{
\item{specdb}{Path to spectra database}

\item{projects}{Character vector of projects for which to load metadata. If 
`NULL` (default), load all projects.}

\item{include_metadata}{Logical. If `TRUE`, include a list column of metadata}

\item{unfactor}{Logical. If `TRUE`, convert factor columns to character}

\item{metadata_cols}{Vector of metadata items to extract as columns}

\item{projects}{Character vector of projects for which to load metadata. If 
`NULL` (default), load all projects.}

\item{include_metadata}{Logical. If `TRUE`, include a list column of metadata}

\item{unfactor}{Logical. If `TRUE`, convert factor columns to character}
}
\value{
Metadata for specified projects, as `tibble`

Metadata for specified projects, as `tibble`
}
\description{
Currently, all metadata is stored in one large list in the object root

Currently, all metadata is stored in one large list in the object root
}
