% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_helpers.R
\name{h5_group}
\alias{h5_group}
\alias{h5_group.H5RefClass}
\alias{h5_group.character}
\title{Open a group, or create it if it doesn't exist}
\usage{
h5_group(hfile, group, overwrite = FALSE)

\method{h5_group}{H5RefClass}(hfile, group, overwrite)

\method{h5_group}{character}(hfile, group, overwrite = FALSE)
}
\arguments{
\item{hfile}{H5File object, or path to HDF5 file.}

\item{group}{Name of group, as character}

\item{overwrite}{Logical. If `TRUE`, unlink the target first.}
}
\description{
Open a group, or create it if it doesn't exist
}
