% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_helpers.R
\name{list2hdf}
\alias{list2hdf}
\title{Recursively insert a list into an HDF5 file}
\usage{
list2hdf(hfile, l, overwrite = FALSE)
}
\arguments{
\item{hfile}{HDF5 file object (see [hdf5r::H5File]).}

\item{l}{A named}

\item{overwrite}{Logical. If `TRUE`, unlink the target first.}
}
\description{
Recursively insert a list into an HDF5 file
}
