% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inversion.R
\name{prepare_inversion}
\alias{prepare_inversion}
\title{Prepare an inversion for submission}
\usage{
prepare_inversion(specdb, prospect_version, inversion_path = "inversion",
  overwrite_all = FALSE)
}
\description{
Prepare an inversion for submission
}
