% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectra.R
\name{read_spectra}
\alias{read_spectra}
\title{Read spectra from HDF5 file}
\usage{
read_spectra(spectra_file, data_name, wavelength = NULL, spectra_id = NULL,
  exact_wl = FALSE, wl_range = NULL)
}
\arguments{
\item{data_name}{Name of data in HDF5 file. If multiple data names are 
passed, spectra are combined via `cbind.spectra`}

\item{wavelength}{Wavelengths to return. If `NULL`, return all wavelengths.}

\item{spectra_id}{Spectra ID to return, as character. If `NULL`, return all spectra}

\item{exact_wl}{Logical, whether or not to parse wavelength specially. See details.}

\item{wl_range}{Numeric, length 2, indicating the minimum and maxiumum 
wavelength to use. Ignored if `wavelength` is not `NULL`.}
}
\description{
Return spectra from a given data set, wavelength (optional), and spectra_id (optional).
}
\details{
If `wavelength` is passed directly as `X:Y`, then all wavelengths between X 
and Y are selected, even if they are not contiguous. For instance, if only 
every 5th wavelength is stored, passing `wavelength = 400:450` will return 
all wavelengths between 400 and 450. This behavior is only triggered if a 
`:` appears in the entered argument -- i.e. this behavior is NOT triggered 
if `wl <- 400:2500; read_spectra(..., wavelength = wl)`. Specifying 
`exact_wl = TRUE` disables this behavior.
}
