% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inversion.R
\name{write_info}
\alias{write_info}
\title{Write log info about a run}
\usage{
write_info(string, file = getOption("inversion_log"), append = TRUE, ...)
}
\arguments{
\item{string}{A string, to be passed to [glue::glue]}

\item{file}{Output file. Defaults to `getOption("inversion_log")` or, if that's unset, the R console}
}
\description{
Write log info about a run
}
